+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * vcf/io/writer/record/info/field/value/integer: Validate value.
+
+    Integer values must now be > -2^31 + 7. See _The Variant Call Format
+    Specification: VCFv4.5 and BCFv2.2_ (2024-06-28) § 1.3 "Data types".
+
 ## 0.65.0 - 2024-09-12
 
 ### Changed